update mangrove species list for 6.2b
