+++ mrtt-ui/src/components/PreRestorationAssessment/PreRestorationAssessmentForm.js
@@ -28,7 +28,7 @@ import {
 import { ContentWrapper } from '../../styles/containers'
 import { ErrorText, PageSubtitle, PageTitle } from '../../styles/typography'
 import { findDataItem } from '../../library/findDataItem'
-import { mangroveSpeciesPerCountryList } from '../../data/mangroveSpeciesPerCountry'
+// import { mangroveSpeciesPerCountryList } from '../../data/mangroveSpeciesPerCountry'
 import { mapDataForApi } from '../../library/mapDataForApi'
 import { multiselectWithOtherValidationNoMinimum } from '../../validation/multiSelectWithOther'
 import { preRestorationAssessment as questions } from '../../data/questions'
@@ -43,6 +43,7 @@ import useInitializeQuestionMappedForm from '../../library/useInitializeQuestion
 import useSiteInfo from '../../library/useSiteInfo'
 import RequiredIndicator from '../RequiredIndicator'
 import FormValidationMessageIfErrors from '../FormValidationMessageIfErrors'
+import organizeMangroveSpeciesList from '../../library/organizeMangroveSpeciesList'
 
 const getSiteCountries = (registrationAnswersFromServer) =>
   findDataItem(registrationAnswersFromServer, '1.2') ?? []
@@ -161,36 +162,9 @@ function PreRestorationAssessmentForm() {
       const siteCountriesResponse = getSiteCountries(serverResponse)
 
       if (siteCountriesResponse.length) {
-        const countriesList = siteCountriesResponse.map(
-          (countryItem) => countryItem.properties.country
-        )
-        // mangroveSpeciesPresent list should display country specific species at the top, with all
-        // species below the country specific list, removing duplicates from the second list
-        const allSpecies = []
-        const countrySelectedSpecies = []
-        let countrySelectedSpeciesWithAllSpecies = []
-        countriesList.forEach((countrySelected) => {
-          mangroveSpeciesPerCountryList.forEach((countryItem) => {
-            if (countryItem.country.name === countrySelected) {
-              countrySelectedSpecies.push(...countryItem.species)
-            }
-            allSpecies.push(...countryItem.species)
-          })
-        })
-        const uniqueCountrySelectedSpecies = [...new Set(countrySelectedSpecies)]
-        uniqueCountrySelectedSpecies.sort()
-        const uniqueAllSpecies = [...new Set(allSpecies)]
-        const filteredUniqueAllSpecies = uniqueAllSpecies.filter(
-          (specie) => !uniqueCountrySelectedSpecies.includes(specie)
-        )
-        filteredUniqueAllSpecies.sort()
+        const organizedSpecies = organizeMangroveSpeciesList(siteCountriesResponse)
 
-        countrySelectedSpeciesWithAllSpecies = [
-          ...uniqueCountrySelectedSpecies,
-          ...filteredUniqueAllSpecies
-        ]
-
-        setMangroveSpeciesList(countrySelectedSpeciesWithAllSpecies)
+        setMangroveSpeciesList(organizedSpecies)
       }
       setMangroveSpeciesTypesChecked(getMangroveSpecies(serverResponse))
