+++ mrtt-ui/src/components/SiteInterventions/SiteInterventionsForm.js
@@ -29,7 +29,6 @@ import { ContentWrapper } from '../../styles/containers'
 import { ErrorText, PageSubtitle, PageTitle } from '../../styles/typography'
 import { findDataItem } from '../../library/findDataItem'
 import { Form, FormPageHeader, FormQuestionDiv, StickyFormLabel } from '../../styles/forms'
-import { mangroveSpeciesPerCountryList } from '../../data/mangroveSpeciesPerCountry'
 import { mapDataForApi } from '../../library/mapDataForApi'
 import { multiselectWithOtherValidationNoMinimum } from '../../validation/multiSelectWithOther'
 import { propaguleOptions, seedlingOptions } from '../../data/siteInterventionOptions'
@@ -44,6 +43,7 @@ import MangroveAssociatedSpeciesRow from './MangroveAssociatedSpeciesRow'
 import QuestionNav from '../QuestionNav'
 import useInitializeQuestionMappedForm from '../../library/useInitializeQuestionMappedForm'
 import useSiteInfo from '../../library/useSiteInfo'
+import organizeMangroveSpeciesList from '../../library/organizeMangroveSpeciesList'
 
 const getWhichStakeholdersInvolved = (registrationAnswersFromServer) =>
   findDataItem(registrationAnswersFromServer, '6.1') ?? []
@@ -166,20 +166,9 @@ function SiteInterventionsForm() {
     const siteCountriesResponse = getSiteCountries(serverResponse)
 
     if (siteCountriesResponse.length) {
-      const countriesList = siteCountriesResponse.map(
-        (countryItem) => countryItem.properties.country
-      )
-      const species = []
-      countriesList.forEach((countrySelected) => {
-        mangroveSpeciesPerCountryList.forEach((countryItem) => {
-          if (countryItem.country.name === countrySelected) {
-            species.push(...countryItem.species)
-          }
-        })
-      })
-      const uniqueSpecies = [...new Set(species)]
+      const organizedSpecies = organizeMangroveSpeciesList(siteCountriesResponse)
 
-      setMangroveSpeciesForCountriesSelected(uniqueSpecies)
+      setMangroveSpeciesForCountriesSelected(organizedSpecies)
     }
 
     const getMangroveSpeciesUsedFrom6_2b = getMangroveSpeciesUsed(serverResponse)
